requirements.txt
tensorflow>=2.13.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.3
PyYAML>=6.0
scikit-learn>=1.1.0
pillow>=8.4.0